+++ packages/shared/src/components/notifications/NotificationsRailPanel.tsx
@@ -14,6 +14,8 @@ import {
   notificationFilterCategoryLabel,
   notificationFilterCategoryList,
 } from './utils';
+import { useConditionalFeature } from '../../hooks/useConditionalFeature';
+import { featureNotificationsRedesign } from '../../lib/featureManagement';
 
 // Compact menu in the rail / v2 context panel. Lists the notification type
 // filters (driven by the `?type=` query param on the notifications page) plus
@@ -27,6 +29,11 @@ export const NotificationsRailPanel = (): ReactElement => {
   const isListPage = router.pathname === '/notifications';
   const activeType =
     typeof router.query?.type === 'string' ? router.query.type : undefined;
+  // Category filters only exist in the redesigned page; the control page
+  // ignores `?type=`, so when the experiment is off keep the simple nav.
+  const { value: isRedesign } = useConditionalFeature({
+    feature: featureNotificationsRedesign,
+  });
 
   // Filters navigate via `action` (button), NOT `path`. SidebarItem treats
   // any `?type=` path as active for the whole `/notifications` route (its
@@ -51,6 +58,41 @@ export const NotificationsRailPanel = (): ReactElement => {
     // Category-owned settings shortcut: keeps the Notifications panel active
     // (the canonical /settings/notifications page keeps the Settings panel).
     const settingsPath = `${webappUrl}notifications/settings`;
+    const allActivityPath = `${webappUrl}notifications`;
+    const unreadBadge = hasUnread && {
+      rightIcon: () => (
+        <Typography
+          type={TypographyType.Caption1}
+          bold
+          className="rounded-6 bg-accent-ketchup-default px-1.5 text-white"
+        >
+          {unreadCount}
+        </Typography>
+      ),
+    };
+
+    const settings: SidebarMenuItem = {
+      title: 'Settings',
+      path: settingsPath,
+      active: isSidebarItemActive(activePage, settingsPath),
+      icon: (active: boolean) => (
+        <ListIcon Icon={() => <SettingsIcon secondary={active} />} />
+      ),
+    };
+
+    if (!isRedesign) {
+      const allActivity: SidebarMenuItem = {
+        title: 'All activity',
+        path: allActivityPath,
+        active: isSidebarItemActive(activePage, allActivityPath),
+        icon: (active: boolean) => (
+          <ListIcon Icon={() => <BellIcon secondary={active} />} />
+        ),
+        ...unreadBadge,
+      };
+
+      return [allActivity, settings];
+    }
 
     const allActivity: SidebarMenuItem = {
       title: 'All activity',
@@ -59,17 +101,7 @@ export const NotificationsRailPanel = (): ReactElement => {
       icon: (active: boolean) => (
         <ListIcon Icon={() => <BellIcon secondary={active} />} />
       ),
-      ...(hasUnread && {
-        rightIcon: () => (
-          <Typography
-            type={TypographyType.Caption1}
-            bold
-            className="rounded-6 bg-accent-ketchup-default px-1.5 text-white"
-          >
-            {unreadCount}
-          </Typography>
-        ),
-      }),
+      ...unreadBadge,
     };
 
     const categoryItems: SidebarMenuItem[] = notificationFilterCategoryList.map(
@@ -86,17 +118,16 @@ export const NotificationsRailPanel = (): ReactElement => {
       },
     );
 
-    const settings: SidebarMenuItem = {
-      title: 'Settings',
-      path: settingsPath,
-      active: isSidebarItemActive(activePage, settingsPath),
-      icon: (active: boolean) => (
-        <ListIcon Icon={() => <SettingsIcon secondary={active} />} />
-      ),
-    };
-
     return [allActivity, ...categoryItems, settings];
-  }, [activePage, activeType, hasUnread, isListPage, navigate, unreadCount]);
+  }, [
+    activePage,
+    activeType,
+    hasUnread,
+    isListPage,
+    isRedesign,
+    navigate,
+    unreadCount,
+  ]);
 
   return (
     <Section
