+++ packages/shared/src/lib/featureManagement.ts
@@ -271,3 +271,11 @@ export const featureDailyPage = new Feature<DailyPageVariant>(
   'daily_page',
   DailyPageVariant.None,
 );
+
+// Experiment: redesigned notifications page (type filters, time grouping,
+// compact rows) backed by server-side type filtering on daily-api. Control is
+// the legacy single-list page. Keep the default `false` — GrowthBook ramps it.
+export const featureNotificationsRedesign = new Feature(
+  'notifications_redesign',
+  false,
+);
